# src/Transfigurr.API.V1/openapi.json
{"openapi": "3.1.0", "info": {"title": "Transfigurr API", "version": "V1"}, "paths": {"/api/season": {"get": {"tags": ["Season"], "summary": "Get All Seasons Route", "operationId": "get_all_seasons_route_api_season_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/season/{season_id}": {"get": {"tags": ["Season"], "summary": "Get Season Route", "operationId": "get_season_route_api_season__season_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "season_id", "in": "path", "required": true, "schema": {"title": "Season Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/season/{season}": {"put": {"tags": ["Season"], "summary": "Set Season Route", "operationId": "set_season_route_api_season__season__put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "season", "in": "path", "required": true, "schema": {"title": "Season"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/scan/series/metadata": {"put": {"tags": ["Scan"], "summary": "Get All Series Metadata Route", "operationId": "get_all_series_metadata_route_api_scan_series_metadata_put", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/scan/series/metadata/{series_id}": {"put": {"tags": ["Scan"], "summary": "Get Series Metadata Route", "operationId": "get_series_metadata_route_api_scan_series_metadata__series_id__put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "series_id", "in": "path", "required": true, "schema": {"title": "Series Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/scan/series": {"put": {"tags": ["Scan"], "summary": "Scan All Series Route", "operationId": "scan_all_series_route_api_scan_series_put", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/scan/series/{series_name}": {"put": {"tags": ["Scan"], "summary": "Scan Series Route", "operationId": "scan_series_route_api_scan_series__series_name__put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "series_name", "in": "path", "required": true, "schema": {"title": "Series Name"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/settings": {"get": {"tags": ["Settings"], "summary": "Get All Settings Route", "operationId": "get_all_settings_route_api_settings_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "put": {"tags": ["Settings"], "summary": "Set Settings", "operationId": "set_settings_api_settings_put", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/settings/{settings_id}": {"get": {"tags": ["Settings"], "summary": "Get Setting Route", "operationId": "get_setting_route_api_settings__settings_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "settings_id", "in": "path", "required": true, "schema": {"title": "Settings Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/codecs": {"get": {"tags": ["Codec"], "summary": "Get All Codecs Route", "operationId": "get_all_codecs_route_api_codecs_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/containers": {"get": {"tags": ["Codec"], "summary": "Get All Containers Route", "operationId": "get_all_containers_route_api_containers_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/encoders": {"get": {"tags": ["Codec"], "summary": "Get All Encoders Route", "operationId": "get_all_encoders_route_api_encoders_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/profiles": {"get": {"tags": ["Profiles"], "summary": "Get All Profiles Route", "operationId": "get_all_profiles_route_api_profiles_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "put": {"tags": ["Profiles"], "summary": "Set Profile Route", "operationId": "set_profile_route_api_profiles_put", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/profiles/{profile_id}": {"get": {"tags": ["Profiles"], "summary": "Get Profile Route", "operationId": "get_profile_route_api_profiles__profile_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "profile_id", "in": "path", "required": true, "schema": {"title": "Profile Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Profiles"], "summary": "Delete Profile Route", "operationId": "delete_profile_route_api_profiles__profile_id__delete", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "profile_id", "in": "path", "required": true, "schema": {"title": "Profile Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/series": {"get": {"tags": ["Series"], "summary": "Get All Series Route", "operationId": "get_all_series_route_api_series_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/series/{series_id}": {"get": {"tags": ["Series"], "summary": "Get Series Route", "operationId": "get_series_route_api_series__series_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "series_id", "in": "path", "required": true, "schema": {"title": "Series Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Series"], "summary": "Set Series Route", "operationId": "set_series_route_api_series__series_id__put", "security": [{"OAuth2PasswordBearer": []}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/system": {"get": {"tags": ["System"], "summary": "Get All System Route", "operationId": "get_all_system_route_api_system_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/system/{system_id}": {"get": {"tags": ["System"], "summary": "Get System Route", "operationId": "get_system_route_api_system__system_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "system_id", "in": "path", "required": true, "schema": {"title": "System Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["System"], "summary": "Set System Route", "operationId": "set_system_route_api_system__system_id__put", "security": [{"OAuth2PasswordBearer": []}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/backdrop/series/{series_id}": {"get": {"tags": ["Artwork"], "summary": "Get Series Backdrop Route", "operationId": "get_series_backdrop_route_api_backdrop_series__series_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "series_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Series Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/poster/series/{series_id}": {"get": {"tags": ["Artwork"], "summary": "Get Series Poster Route", "operationId": "get_series_poster_route_api_poster_series__series_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "series_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Series Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/history": {"get": {"tags": ["History"], "summary": "Get All History Route", "operationId": "get_all_history_route_api_history_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "put": {"tags": ["History"], "summary": "Set Settings", "operationId": "set_settings_api_history_put", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/history/{history_id}": {"get": {"tags": ["History"], "summary": "Get Setting", "operationId": "get_setting_api_history__history_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "history_id", "in": "path", "required": true, "schema": {"title": "History Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/episode": {"get": {"tags": ["Episode"], "summary": "Get All Episodes Route", "operationId": "get_all_episodes_route_api_episode_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/episode/{episode_id}": {"get": {"tags": ["Episode"], "summary": "Get Episode Route", "operationId": "get_episode_route_api_episode__episode_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "episode_id", "in": "path", "required": true, "schema": {"title": "Episode Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Episode"], "summary": "Set Episode Route", "operationId": "set_episode_route_api_episode__episode_id__put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "episode_id", "in": "path", "required": true, "schema": {"title": "Episode Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/login": {"post": {"tags": ["Authentication"], "summary": "Login Route", "operationId": "login_route_api_login_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/register": {"post": {"tags": ["Authentication"], "summary": "Register Route", "operationId": "register_route_api_register_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/activated": {"get": {"tags": ["Authentication"], "summary": "Get Activated Route", "operationId": "get_activated_route_api_activated_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/logintoken": {"post": {"tags": ["Authentication"], "summary": "Login With Token Route", "operationId": "login_with_token_route_api_logintoken_post", "parameters": [{"name": "authorization", "in": "header", "required": false, "schema": {"type": "string", "title": "Authorization"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/actions/restart": {"put": {"tags": ["Actions"], "summary": "Restart Route", "operationId": "restart_route_api_actions_restart_put", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/actions/shutdown": {"put": {"tags": ["Actions"], "summary": "Shutdown Route", "operationId": "shutdown_route_api_actions_shutdown_put", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/api/user": {"get": {"tags": ["User"], "summary": "Get User Route", "operationId": "get_user_route_api_user_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "put": {"tags": ["User"], "summary": "Set Settings", "operationId": "set_settings_api_user_put", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/{full_path}": {"get": {"summary": "Read Item", "operationId": "read_item__full_path__get", "parameters": [{"name": "full_path", "in": "path", "required": true, "schema": {"type": "string", "title": "Full Path"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "login"}}}}}}